[Bug] fix(ios): resolve web compilation error and lipo thin-payload error

+++ lib/utils/remote_media_fetcher.dart
@@ -39,7 +39,7 @@ class RemoteMediaFetcher {
         );
         return RemoteMediaHead(
           fileName: result.fileName,
-          fileSize: result.fileSize,
+          fileSize: result.fileSize.toInt(),
           headBytes: Uint8List(0),
           bytesHashed: result.bytesHashed,
           hash: result.hash,
